Consolidate the ItSimpleDomainValidation.java ItSimpleOperatorValidation.java and ItWitValidation.java into a single file
